# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Job Board API
  version: 1.0.0
  description: API for managing job listings

servers:
  - url: http://localhost:3000/api
    description: Development server

paths:
  /jobs:
    get:
      summary: Get all jobs
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      summary: Create a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

  /jobs/{id}:
    get:
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found

components:
  schemas:
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        description:
          type: string
        location:
          type: string
        application_method:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - title
        - company
        - description
        - location
        - application_method

    JobInput:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        description:
          type: string
        location:
          type: string
        application_method:
          type: string
      required:
        - title
        - company
        - description
        - location
        - application_method